requirements.txt
flask==3.0.3
python-dotenv==1.0.1
langchain==0.3.1
langchain-community==0.3.1
langchain-google-genai==2.0.10
